move api models to library
