+++ biosimulations/libs/dispatch/api-models/src/index.ts
@@ -1,2 +1,4 @@
 export * from './lib/dispatch-api-models';
 export * from './lib/common/dispatch-simulation.model';
+export * from './lib/simulationRun';
+export * from './lib/simulationRunStatus';
